% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nwka.R
\name{read_nwka_nexus}
\alias{read_nwka_nexus}
\title{Read Tree File in NEXUS Format with NWKA Trees}
\usage{
read_nwka_nexus(file, tree.names = NULL, force.multi = FALSE, debug = FALSE)
}
\arguments{
\item{file}{A file name.}

\item{tree.names}{A vector of mode character containing names for the trees that are read from the file;
if \code{NULL} (the default), the trees will be named according to the names in the tree file or,
if these are missing, as \code{"tree1"}, \code{"tree2"}, ...}

\item{force.multi}{If \code{TRUE}, this function will return an object of class \code{"multiPhylo"} even
when the tree file contains only a single tree. Defaults to \code{FALSE}, which means that if the
file contains a single tree, an object of class \code{"phylo"} is returned.}

\item{debug}{A logical value indicating whether to enable verbose debug output while parsing the tree. If this is \code{TRUE},
the function will print information about each node in the each tree as it parses it.}
}
\value{
An object of class \code{"phylo"} or \code{"multiPhylo"}, compatible with the \code{\link[ape]{ape}}
        package.

        In addition to the elements described in the documentation for the \code{\link[ape]{read.tree}}
        function of the \code{\link[ape]{ape}} package, a \code{"phylo"} object produced by this function
        will also have the following components:
        \item{\code{tip.attributes}}{A named list of attributes for the tips of the tree. Each element of
                                     this list is a vector of mode character or numeric (depending on the attribute).}
        \item{\code{node.attributes}}{A named list of attributes for the internal nodes of the tree. Each element of
                                      this list is a vector of mode character or numeric (depending on the attribute).}
}
\description{
This function reads a file containing one or more trees in NEXUS format. Each tree is parsed according to the Newick-with-Attributes (NWKA) format.
}
\details{
Only the \code{Trees} block of the NEXUS file is parsed.

         Node attributes (e.g. support values, rates, ages...) are parsed by this function and returned in the
         \code{tip.attributes} and \code{node.attributes} elements of the returned \code{"phylo"} objects. If the nodes
         contain a \code{prob} attribute, its value will also be copied to the \code{Support} attribute.

         The translation table (if any) of the \code{Trees} block is used to translate the names of both tips and internal
         nodes. However, if the untranslated names of internal nodes are numbers, these may be interpreted as support
         values (and thus, not translated).

         Attribute names may appear in any kind of casing (e.g. \code{Name}, \code{name} or \code{NAME}), but they
         should be treated using case-insensitive comparisons.

         Setting the \code{debug} argument to \code{TRUE} can be useful when analysing malformed trees (to understand
         at which point in the tree the problem lies).
}
\references{
\url{https://github.com/arklumpus/TreeNode/blob/master/NWKA.md}
}
\seealso{
\code{\link[ape]{ape}}, \code{\link[ape]{read.tree}}, \code{\link[ape]{read.nexus}}

Other functions to read trees: 
\code{\link{read_binary_trees}()},
\code{\link{read_nwka_tree}()},
\code{\link{read_one_binary_tree}()}
}
\author{
Giorgio Bianchini
}
\concept{functions to read trees}
