% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nwka.R
\name{read_nwka_tree}
\alias{read_nwka_tree}
\title{Read Tree File in NWKA Format}
\usage{
read_nwka_tree(
  file = "",
  text = NULL,
  tree.names = NULL,
  keep.multi = FALSE,
  debug = FALSE
)
}
\arguments{
\item{file}{A file name.}

\item{text}{A variable of mode character containing the tree(s) to parse. By default, this is set to \code{NULL} and
ignored (i.e. the tree is read from the file specified by the \code{file} argument); otherwise, the file argument
is ignored and the trees are read from the \code{text} argument.}

\item{tree.names}{A vector of mode character containing names for the trees that are read from the file;
if \code{NULL} (the default), the trees will be named as \code{"tree1"}, \code{"tree2"}, ...}

\item{keep.multi}{If \code{TRUE}, this function will return an object of class \code{"multiPhylo"} even
when the tree file contains only a single tree. Defaults to \code{FALSE}, which means that if the
file contains a single tree, an object of class \code{"phylo"} is returned.}

\item{debug}{A logical value indicating whether to enable verbose debug output while parsing the tree. If this is \code{TRUE},
the function will print information about each node in the tree as it parses it.}
}
\value{
An object of class \code{"phylo"} or \code{"multiPhylo"}, compatible with the \code{\link[ape]{ape}}
        package.

        In addition to the elements described in the documentation for the \code{\link[ape]{read.tree}}
        function of the \code{\link[ape]{ape}} package, a \code{"phylo"} object produced by this function
        will also have the following components:
        \item{\code{tip.attributes}}{A named list of attributes for the tips of the tree. Each element of
                                     this list is a vector of mode character or numeric (depending on the attribute).}
        \item{\code{node.attributes}}{A named list of attributes for the internal nodes of the tree. Each element of
                                      this list is a vector of mode character or numeric (depending on the attribute).}
}
\description{
This function reads a file containing one or more trees in Newick-with-Attributes (NWKA) format.
}
\details{
The Newick-with-Attributes format parsed by this function is backwards compatible with the Newick/New Hampshire
         format and some of its extensions (e.g. Extended Newick, New Hampshire X).

         Node attributes (e.g. support values, rates, ages...) are parsed by this function and returned in the
         \code{tip.attributes} and \code{node.attributes} elements of the returned \code{"phylo"} objects. If the nodes
         contain a \code{prob} attribute, its value will also be copied to the \code{Support} attribute.

         Attribute names may appear in any kind of casing (e.g. \code{Name}, \code{name} or \code{NAME}), but they
         should be treated using case-insensitive comparisons.

         Setting the \code{debug} argument to \code{TRUE} can be useful when analysing malformed trees (to understand
         at which point in the tree the problem lies).
}
\examples{
# Parse a tree string
# Topology from https://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?id=207598
tree <- read_nwka_tree(text="(((('Homo sapiens'[rank=species])'Homo'[rank=genus])
'Hominina'[rank=subtribe],(('Pan paniscus'[rank=species],'Pan troglodytes'
[rank=species])'Pan'[rank=genus])'Panina'[rank=subtribe])'Hominini'[rank=tribe],
(('Gorilla gorilla'[rank=species],'Gorilla beringei'[rank=species])'Gorilla'
[rank=genus])'Gorillini'[rank=tribe])'Homininae'[rank=subfamily];")

# Show the tree's structure
str(tree)

# Plot the tree with node labels
ape::plot.phylo(tree, show.node.label = TRUE, node.depth = 2)

# Add taxonomic rank (stored in the "rank" attribute of the tree)
tree$tip.label = paste(tree$tip.attributes$rank, tree$tip.attributes$Name, sep="\n")
tree$node.label = paste(tree$node.attributes$rank, tree$node.attributes$Name, sep="\n")

# Plot again
ape::plot.phylo(tree, show.node.label = TRUE, node.depth = 2, y.lim=c(0.5, 5.5))

}
\references{
\url{https://github.com/arklumpus/TreeNode/blob/master/NWKA.md}
}
\seealso{
\code{\link[ape]{ape}}, \code{\link[ape]{read.tree}}

Other functions to read trees: 
\code{\link{read_binary_trees}()},
\code{\link{read_nwka_nexus}()},
\code{\link{read_one_binary_tree}()}
}
\author{
Giorgio Bianchini
}
\concept{functions to read trees}
