% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_binary_trees.R
\name{read_one_binary_tree}
\alias{read_one_binary_tree}
\title{Read Tree in Binary Format}
\usage{
read_one_binary_tree(file, index = 1, address = NA, metadata = NA)
}
\arguments{
\item{file}{A file name.}

\item{index}{The index of the tree that should be read (starting from 1).}

\item{address}{The address (i.e. byte offset from the start of the file) of the tree that should be read.}

\item{metadata}{An object of class \code{"BinaryTreeMetadata"} containing the metadata extracted from the
tree file. If this is not provided, it will be read from the file (see details).}
}
\value{
An object of class \code{"phylo"}, compatible with the \code{\link[ape]{ape}}
        package.

        In addition to the elements described in the documentation for the \code{\link[ape]{read.tree}}
        function of the \code{\link[ape]{ape}} package, a \code{"phylo"} object produced by this function
        will also have the following components:
        \item{\code{tip.attributes}}{A named list of attributes for the tips of the tree. Each element of
                                     this list is a vector of mode character or numeric (depending on the attribute).}
        \item{\code{node.attributes}}{A named list of attributes for the internal nodes of the tree. Each element of
                                      this list is a vector of mode character or numeric (depending on the attribute).}
}
\description{
This function reads one tree from a file in binary format.
}
\details{
This function extracts only one tree from the file. The information provided by \code{metadata} is used to
         determine where in the file the requested tree starts. If this is not provided, this function will read the
         metadata from the file (using the \code{\link{read_binary_tree_metadata}} function).

         Reusing the metadata is efficient when multiple trees need to be read from the same file (so that the metadata
         only needs to be read once).

         Node attributes (e.g. support values, rates, ages...) are parsed by this function and returned in the
         \code{tip.attributes} and \code{node.attributes} elements of the returned \code{"phylo"} objects.

         Attribute names may appear in any kind of casing (e.g. \code{Name}, \code{name} or \code{NAME}), but they
         should be treated using case-insensitive comparisons.

         Due to limitations with R's integral types, this function may have issues with files larger than 2GB.
}
\examples{
# Tree file (replace with your own)
treeFile <- system.file("extdata", "manyTrees.tbi", package="TreeNode")

# Read the 5th tree in the file
tree <- read_one_binary_tree(treeFile, 5)
#Do something with the tree


# Read the binary tree metadata
meta <- read_binary_tree_metadata(treeFile)

# Process every tree in the file
for (add in meta$TreeAddresses)
{
    tree <- read_one_binary_tree(treeFile, address = add, metadata = meta)
    #Do something with the tree
}


}
\references{
\url{https://github.com/arklumpus/TreeNode/blob/master/BinaryTree.md}
}
\seealso{
\code{\link{read_binary_trees}}, \code{\link{read_binary_tree_metadata}}, \code{\link[ape]{ape}}, \code{\link[ape]{read.tree}}

Other functions to read trees: 
\code{\link{read_binary_trees}()},
\code{\link{read_nwka_nexus}()},
\code{\link{read_nwka_tree}()}
}
\author{
Giorgio Bianchini
}
\concept{functions to read trees}
