% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeNode.R
\docType{package}
\name{TreeNode}
\alias{TreeNode}
\title{TreeNode: Read and Write Trees in Binary and NWKA Formats}
\description{
The \strong{TreeNode} package provides functions to read and write files containing
phylogenetic trees in Binary Tree format and in Newick-with-Attributes (NWKA)
format.

\strong{TreeNode} produces and consumes trees stored in the same \code{"phylo"} or
\code{"multiPhylo"} objects used by the package \link[ape]{ape}.

More information on \strong{TreeNode} can be found at \url{https://github.com/arklumpus/TreeNode}.
}
\references{
Paradis, E., Claude, J. and Strimmer, K. (2004) \emph{APE: analyses of phylogenetics and evolution in R language.} Bioinformatics, \strong{20}, 289-290
}
\seealso{
\url{https://github.com/arklumpus/TreeNode}, \code{\link[ape]{ape}}
}
\author{
Giorgio Bianchini

\strong{Maintainer}: Giorgio Bianchini <giorgio.bianchini@bristol.ac.uk>
}
