% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_binary_trees.R
\name{write_binary_trees}
\alias{write_binary_trees}
\title{Write Tree File in Binary Format}
\usage{
write_binary_trees(trees, file, additional_data = vector("raw", 0))
}
\arguments{
\item{trees}{An object of class \code{"phylo"} or \code{"multiPhylo"}.}

\item{file}{A file name.}

\item{additional_data}{A vector of mode raw containg additional binary data that will be included within
the tree file.}
}
\description{
This function writes one or more trees to a file in binary format.
}
\details{
This function writes all the trees at once. If you wish to write the trees one at a time, you
         should use the \code{\link{keep_writing_binary_trees}} function.

         This function will analyse all the trees to determine whether it is appropriate to include any
         names or attributes in the file header. It will then write the header, the trees and conclude the
         file with an appropriate trailer.

         The tip names can be specified either as a \code{Name} element in the tree's \code{tip.attributes}
         element, or as the \code{tip.label} element of the tree. If both are specified, the values stored in
         the \code{Name} attribute take precedence (this allows backward compatibility for trees created
         using \code{\link[ape]{ape}}).

         The node names and support values can similarly be specified either with a \code{Name} or \code{Support}
         element in the tree's \code{node.attributes}, or as the tree's \code{node.label}. If all the node labels
         can be parsed as numbers, they will be assumed to contain support values; otherwise, they will be
         assumed to contain node names. If the \code{node.attributes} already contain a \code{Name} or \code{Support}
         element, the node labels will be ignored.

         The additional binary data (if any) will be written in the file after the trees and before the trailer.
}
\references{
\url{https://github.com/arklumpus/TreeNode/blob/master/BinaryTree.md}
}
\seealso{
\code{\link{keep_writing_binary_trees}}, \code{\link[ape]{ape}}, \code{\link[ape]{write.tree}}

Other functions to write trees: 
\code{\link{keep_writing_binary_trees}()},
\code{\link{write_nwka_nexus}()},
\code{\link{write_nwka_tree}()}
}
\author{
Giorgio Bianchini
}
\concept{functions to write trees}
