% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nwka.R
\name{write_nwka_tree}
\alias{write_nwka_tree}
\title{Write Tree File in NWKA format}
\usage{
write_nwka_tree(trees, file = "", append = FALSE, nwka = TRUE, quotes = FALSE)
}
\arguments{
\item{trees}{An object of class \code{"phylo"} or \code{"multiPhylo"}.}

\item{file}{A file name. If this is \code{""} (the default), the tree will be written on the standard
output.}

\item{append}{If this is \code{FALSE} (the default), the output file (if it exists already) is truncated
before writing trees (i.e. overwritten). If this is \code{TRUE}, the trees are appended at
the end of the output file.}

\item{nwka}{If this is \code{TRUE} (the default), the tree will be written in Newick-with-Attributes
(NWKA) format. Otherwise, the tree will be written in Newick format (and attributes that
cannot be represented in this format will be lost).}

\item{quotes}{If \code{nwka = FALSE}, this argument determines whether names in the tree file will be
enclosed within single quotes (if this is \code{TRUE}) or not (if this is \code{FALSE}).}
}
\description{
This function writes one or more trees in Newick-with-Attributes (NWKA) format to a file or to the standard output.
}
\details{
All of the available attributes are written to the file if \code{nwka = TRUE}. Otherwise, (if
         available) the tip names and lenghts are always written, as well as the internal nodes' lenghts and support
         values. If nodes have a name, this is only included if they do not have a support value as well.

         The tip names can be specified either as a \code{Name} element in the tree's \code{tip.attributes}
         element, or as the \code{tip.label} element of the tree. If both are specified, the values stored in
         the \code{Name} attribute take precedence (this allows backward compatibility for trees created
         using \code{\link[ape]{ape}}).

         The node names and support values can similarly be specified either with a \code{Name} or \code{Support}
         element in the tree's \code{node.attributes}, or as the tree's \code{node.label}. If all the node labels
         can be parsed as numbers, they will be assumed to contain support values; otherwise, they will be
         assumed to contain node names. If the \code{node.attributes} already contain a \code{Name} or \code{Support}
         element, the node labels will be ignored.

         No attempt is made to fix problematic labels. Thus, if the tip or node names contain special characters,
         an invalid output may be produced. For example, if the labels contain spaces or commas and \code{enwk} and
         \code{quotes} are both \code{FALSE}, the output tree may not be parsed correctly. If you wish to produce a
         tree conforming to the Newick format while fixing problematic tip labels, you should look into the
         \code{\link[ape]{write.tree}} function of the \code{\link[ape]{ape}} package.
}
\examples{
# Parse a tree string
# Topology from https://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?id=207598
tree <- read_nwka_tree(text="(((('Homo sapiens'[rank=species])'Homo'[rank=genus])
'Hominina'[rank=subtribe],(('Pan paniscus'[rank=species],'Pan troglodytes'
[rank=species])'Pan'[rank=genus])'Panina'[rank=subtribe])'Hominini'[rank=tribe],
(('Gorilla gorilla'[rank=species],'Gorilla beringei'[rank=species])'Gorilla'
[rank=genus])'Gorillini'[rank=tribe])'Homininae'[rank=subfamily];")

# Print the tree to the standard output in NWKA format
cat(write_nwka_tree(tree))

# Print the tree to the standard output in Newick format without quotes
cat(write_nwka_tree(tree, nwka = FALSE))

# Print the tree to the standard output in Newick format with quotes
cat(write_nwka_tree(tree, nwka = FALSE, quotes = TRUE))

}
\references{
\url{https://github.com/arklumpus/TreeNode/blob/master/NWKA.md}
}
\seealso{
\code{\link[ape]{ape}}, \code{\link[ape]{write.tree}}

Other functions to write trees: 
\code{\link{keep_writing_binary_trees}()},
\code{\link{write_binary_trees}()},
\code{\link{write_nwka_nexus}()}
}
\author{
Giorgio Bianchini
}
\concept{functions to write trees}
